Fast_api/requirements.txt
FastAPI-CLI==0.0.5
fastapi==0.112.2
uvicorn==0.30.6
mysql-connector-python==9.0.0
SQLAlchemy==2.0.31
python-dotenv==1.0.1
firebase-admin==6.5.0
requests==2.32.3
PyJWT==2.9.0
psycopg2-binary==2.9.10